app/requirements.txt
streamlit>=1.32.0
pandas>=1.5.3
snowflake-connector-python>=3.7.0
plotly>=5.18.0
python-dotenv>=1.0.0
altair>=5.2.0
matplotlib>=3.8.0